Support for ColumnAttribute
It is now possible to set the column name of each property, by using the
Column Attribute.
There are also some minor fixes (such as insert returns the Identity),
but since the tests use sqlce, it has no support for the get_identity
function. the tests cannot be run :( (however, they didn't compile in
the first place)