Fix returning wrong identity GUID on insert
A GUID is being created and returned on insert but never actually
applied to the insert query. This can be seen on a table where the key
is a GUID and defaults to (newid()) and no GUID is provided on the
insert. Dapper will generate a GUID but it is not applied so the GUID is
generated by newid() but the Dapper GUID is returned instead which is
incorrect.