Rewrote update and insert to use generics. NOTE: this is a breaking change!
Updates and inserts now look at the properties of the type passed in or specified as a generic. Previous functionality can be preserved by passing in with type 'object'.
Previously, since it was just using objects, GetScaffoldableProperties would get properties of the object, rather than the type it was being used as.
This was causing issues if you have an instance of Foo where Foo : Bar and passed to simplecrud with Foo as Bar.
Instead of only using Bars properties, it would look at Foo's too

Squashed commits:
removed now unnecessary Activator.CreateInstance calls (+1 squashed commits)
Added generic versions of insert and update - the most impactful change was to GetScaffoldableProperties. Existing functionality of Update has been modified.